some refactoring.
