+++ packages/story-editor/src/components/colorPicker/colorAdd.js
@@ -27,7 +27,7 @@ import { Icons } from '@web-stories-wp/design-system';
  */
 import { focusStyle } from '../panels/shared';
 import useAddPreset from '../panels/design/preset/useAddPreset';
-import { PRESET_TYPES } from '../panels/design/preset/constants';
+import { PRESET_TYPES } from '../../constants';
 
 const COLOR_SIZE = 32;
 const AddColorAction = styled.button`
