+++ packages/story-editor/src/components/panels/design/preset/presetPanel.js
@@ -30,7 +30,7 @@ import { __ } from '@web-stories-wp/i18n';
  */
 import { useStory } from '../../../../app/story';
 import { Panel, PanelContent } from '../../panel';
-import { areAllType, getPanelInitialHeight } from './utils';
+import { areAllType, getPanelInitialHeight } from '../../../../utils/presetUtils';
 import PresetsHeader from './header';
 import Resize from './resize';
 import useAddPreset from './useAddPreset';
@@ -38,7 +38,7 @@ import useDeletePreset from './useDeletePreset';
 import StyleGroup from './stylePreset/styleGroup';
 import useApplyStyle from './stylePreset/useApplyStyle';
 import ConfirmationDialog from './confirmationDialog';
-import { PRESET_TYPES } from './constants';
+import { PRESET_TYPES } from '../../../../constants';
 
 function PresetPanel({ pushUpdate }) {
   const { selectedElements, globalStoryStyles } = useStory(
