+++ packages/story-editor/src/components/panels/design/preset/test/stylePresetPanel.js
@@ -22,13 +22,16 @@ import { fireEvent, screen } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import StylePresetPanel from '../stylePreset';
+import StylePresetPanel from '../presetPanel';
 import StoryContext from '../../../../../app/story/context';
-import { BACKGROUND_TEXT_MODE } from '../../../../../constants';
-import { getTextPresets, areAllType, getPanelInitialHeight } from '../utils';
+import { BACKGROUND_TEXT_MODE , PRESET_TYPES } from '../../../../../constants';
+import {
+  getTextPresets,
+  areAllType,
+  getPanelInitialHeight,
+} from '../../../../../utils/presetUtils';
 import { renderWithTheme } from '../../../../../testUtils';
 import { TEXT_ELEMENT_DEFAULT_FONT } from '../../../../../app/font/defaultFonts';
-import { PRESET_TYPES } from '../constants';
 
 jest.mock('../utils');
